+++ horovod/ray/runner.py
@@ -10,6 +10,7 @@ import logging
 
 from horovod.runner.common.util import secret, timeout, hosts
 from horovod.runner.http.http_server import RendezvousServer
+from horovod.ray import ray_logger
 from horovod.ray.utils import detect_nics, nics_to_env_var
 
 logger = logging.getLogger(__name__)
@@ -89,6 +90,10 @@ class BaseHorovodWorker:
     def execute(self, func):
         """Executes an arbitrary function on self."""
         return func(self.executable)
+
+    def set_queue(self, queue):
+        """Sets the queue for multi-node logging."""
+        ray_logger.configure(queue=queue)
 
 
 @ray.remote